+++ tests/Common/CommonCommandTest.php
@@ -306,7 +306,7 @@ abstract class CommonCommandTest extends AbstractCommandTest
     public function testBatchInsert(
         string $table,
         array $columns,
-        array $values,
+        iterable $values,
         string $expected,
         array $expectedParams = [],
         int $insertedRow = 1